Move WindowSizeData to script_traits.
