+++ components/compositing/compositor.rs
@@ -28,16 +28,16 @@ use layers::platform::surface::NativeDisplay;
 use layers::rendergl;
 use layers::rendergl::RenderContext;
 use layers::scene::Scene;
-use msg::constellation_msg::{Image, PixelFormat};
-use msg::constellation_msg::{Key, KeyModifiers, KeyState, LoadData};
-use msg::constellation_msg::{NavigationDirection, PipelineId, PipelineIndex, PipelineNamespaceId};
-use msg::constellation_msg::{WindowSizeData, WindowSizeType};
+use msg::constellation_msg::{Image, PixelFormat, Key, KeyModifiers, KeyState};
+use msg::constellation_msg::{LoadData, NavigationDirection, PipelineId};
+use msg::constellation_msg::{PipelineIndex, PipelineNamespaceId, WindowSizeType};
 use profile_traits::mem::{self, ReportKind, Reporter, ReporterRequest};
 use profile_traits::time::{self, ProfilerCategory, profile};
 use script_traits::CompositorEvent::{MouseMoveEvent, MouseButtonEvent, TouchEvent};
 use script_traits::{AnimationState, AnimationTickType, ConstellationControlMsg};
 use script_traits::{ConstellationMsg, LayoutControlMsg, MouseButton, MouseEventType};
-use script_traits::{StackingContextScrollState, TouchpadPressurePhase, TouchEventType, TouchId};
+use script_traits::{StackingContextScrollState, TouchpadPressurePhase, TouchEventType};
+use script_traits::{TouchId, WindowSizeData};
 use std::collections::hash_map::Entry::{Occupied, Vacant};
 use std::collections::{HashMap, HashSet};
 use std::fs::File;
