+++ components/msg/Cargo.toml
@@ -17,13 +17,11 @@ heapsize = "0.3.0"
 heapsize_plugin = "0.1.2"
 hyper = {version = "0.9.9", features = ["serde-serialization"]}
 ipc-channel = {git = "https://github.com/servo/ipc-channel"}
-layers = {git = "https://github.com/servo/rust-layers", features = ["plugins"]}
 plugins = {path = "../plugins"}
 rustc-serialize = "0.3.4"
 serde = "0.7.11"
 serde_macros = "0.7.11"
 url = {version = "1.0.0", features = ["heap_size", "serde"]}
-util = {path = "../util", features = ["servo"]}
 
 [dependencies.webrender_traits]
 git = "https://github.com/servo/webrender_traits"
