+++ components/msg/lib.rs
@@ -15,11 +15,9 @@ extern crate euclid;
 extern crate heapsize;
 extern crate hyper;
 extern crate ipc_channel;
-extern crate layers;
 extern crate rustc_serialize;
 extern crate serde;
 extern crate url;
-extern crate util;
 extern crate webrender_traits;
 
 pub mod constellation_msg;
